refactor(transform): extract attribute_converter.go below the line cap (#375)
attribute_converter.go sat at 490/500 lines. Split it along two seams:

- relationRootsFor and checkRequiredAttributes move to attribute_required.go,
  next to the isRequiredAttributeMissing helpers they are the only caller of,
  so the whole required policy lives in one file.
- The receiver-free scalar coercion helpers (toTimeForEAV, toBoolForEAV,
  derefPointer, toTime, toBool, ToFloat64 and friends) move to a new
  scalar_conversion.go, mirroring numeric_conversion.go.

attribute_converter.go now holds only the converter type and its four
conversion methods (213 lines).

While here, wrap the two bare `return nil, err` sites in FromEAVRecords
deferred from the #373 review: the GetSchemaMetadata and
checkRequiredAttributes failures now carry the operation and schema ID.
Both stay plain errors; every assertion on these messages is a
strings.Contains, so the prefix changes no test.
